Suspend Invocation during class loading
This fixes test flakiness; particularly tests succeeding when we're
running the whole suite, but failing when ran individually.

Removing the workaround in FlowTester which is no longer needed now.

This does not fix the problem yet for class initialization (<clinit>).

This, in combination with tracking of extra args which we did just
before, fixes StringTest.testReplaceChars/testReplaceSingleCharString.