Improve development experience more with flowtracker-agent-dev
Make DevAgent a subclass of FlowTrackAgent. Cleaner, and avoids needing
other tricks to find out if we're doing the dev thing or not.

Include external dependencies in the flowtracker-agent-dev jar (just
like in flowtracker-all), and automatically include flowtracker-weaver
and flowtracker-web classes on the spider classpath.
That removes the need for passing in the spiderClasspath as agent arg.