Add SourceResource.Line.parts
Let the server return the mapping from lines in the source code, to
which regions in the tracker that corresponds to.
That simplifies the work in the UI: we can now directly give a selection
as it was selected in a sink, and from that determine which lines that
corresponds to, without the UI having to calculate that through the
line mapping in the regions of the ClassOriginTracker's
TrackerDetailResponse.